Refactor enrichLpLeaderboards fn
