+++ apps/veil/src/entities/leaderboard/api/use-lp-leaderboard.ts
@@ -1,48 +1,18 @@
 import { useQuery, UseQueryResult } from '@tanstack/react-query';
-import { Position } from '@penumbra-zone/protobuf/penumbra/core/component/dex/v1/dex_pb';
 import { apiPostFetch } from '@/shared/utils/api-fetch';
+import { statusStore } from '@/shared/model/status';
+import { useCurrentEpoch } from '@/pages/tournament/api/use-current-epoch';
 import {
   LpLeaderboardRequest,
   LpLeaderboardApiResponse,
   LpLeaderboardSortKey,
   LpLeaderboardSortDirection,
-  LqtLp,
-  LpLeaderboard,
   LpLeaderboardResponse,
-} from '@/entities/leaderboard/api/utils';
-import { penumbra } from '@/shared/const/penumbra';
-import { createClient } from '@connectrpc/connect';
-import { DexService } from '@penumbra-zone/protobuf';
-import { bech32mPositionId } from '@penumbra-zone/bech32m/plpid';
-import { statusStore } from '@/shared/model/status';
-import { useCurrentEpoch } from '@/pages/tournament/api/use-current-epoch';
-import { connectionStore } from '@/shared/model/connection';
-import { getGrpcTransport } from '@/shared/api/transport';
+  enrichLpLeaderboards,
+} from './utils';
 
 export const BASE_LIMIT = 10;
 export const BASE_PAGE = 1;
-
-// get the position state for each lp reward
-async function enrichLpLeaderboards(data: LqtLp[]): Promise<LpLeaderboard[]> {
-  if (data.length === 0) {
-    return [];
-  }
-
-  const grpc = !connectionStore.connected ? await getGrpcTransport() : undefined;
-  const payload = { positionId: data.map(lp => lp.positionId) };
-  const positionsRes = await Array.fromAsync(
-    grpc
-      ? createClient(DexService, grpc.transport).liquidityPositionsById(payload)
-      : penumbra.service(DexService).liquidityPositionsById(payload),
-  );
-  const positions = positionsRes.map(r => r.data).filter(Boolean) as Position[];
-
-  return data.map((lp, index) => ({
-    ...lp,
-    positionIdString: bech32mPositionId(lp.positionId),
-    position: positions[index] as unknown as Position,
-  }));
-}
 
 export const useLpLeaderboard = ({
   epoch,
