+++ apps/veil/src/entities/leaderboard/api/utils.ts
@@ -4,6 +4,12 @@ import {
   Position,
   PositionId,
 } from '@penumbra-zone/protobuf/penumbra/core/component/dex/v1/dex_pb';
+import { connectionStore } from '@/shared/model/connection';
+import { getGrpcTransport } from '@/shared/api/transport';
+import { penumbra } from '@/shared/const/penumbra';
+import { createClient } from '@connectrpc/connect';
+import { DexService } from '@penumbra-zone/protobuf';
+import { bech32mPositionId } from '@penumbra-zone/bech32m/plpid';
 
 const SORT_KEYS = ['executions', 'points'] as const;
 export type LpLeaderboardSortKey = (typeof SORT_KEYS)[number];
@@ -47,4 +53,26 @@ export interface LpLeaderboard extends LqtLp {
 export interface LpLeaderboardResponse extends LpLeaderboardApiResponse {
   data: LpLeaderboard[];
   total: number;
+}
+
+// get the position state for each lp reward
+export async function enrichLpLeaderboards(data: LqtLp[]): Promise<LpLeaderboard[]> {
+  if (data.length === 0) {
+    return [];
+  }
+
+  const grpc = !connectionStore.connected ? await getGrpcTransport() : undefined;
+  const payload = { positionId: data.map(lp => lp.positionId) };
+  const positionsRes = await Array.fromAsync(
+    grpc
+      ? createClient(DexService, grpc.transport).liquidityPositionsById(payload)
+      : penumbra.service(DexService).liquidityPositionsById(payload),
+  );
+  const positions = positionsRes.map(r => r.data).filter(Boolean) as Position[];
+
+  return data.map((lp, index) => ({
+    ...lp,
+    positionIdString: bech32mPositionId(lp.positionId),
+    position: positions[index] as unknown as Position,
+  }));
 }
